// frontend/src/app/core/models/Autenticacion/DatosUsuarioLogin.ts
export class DatosUsuarioLogin {
  tipoPersona: string;
  descTipoPersona: string;
  nroDocumento: string;
  tipoDocumento: string;
  razonSocial: string;
  nombres: string;
  apePaterno: string;
  apeMaterno: string;
  direccion: string;
  ubigeo: string;
  departamento: string;
  provincia: string;
  distrito: string;
  correo: string;
  telefono: string;
  celular: string;
  ruc: string

  tipoPersonaDesc: string;
  tipoDocumentoDesc: string;

  nombreCompleto: string;
}